Adds the ability to select your default boss kit as well as an item that opens the menu
Various boss changes, as well as adding a new boss the wither

Fixes fish flurry sounds being incorrect and it continuing to shoot projectiles even without an owner

Fixes wither image targeting dead players and an exception when the owner is cleared

Fixes lightning shield effects targeting a different damager if the event damager is changed later on (say, displaying and attributing image damage to the original owner by changing the damager to the owner)

NoVoidDeath now spawns the person at a proper respawn point

Bosses can now specify their health per player with a new interface BossKitData

Adds explosions to the block restore manager

The block restore manager now stops all falling blocks from forming

Damage multipliers on kits now cap out at 0 instead of going negative

Scoreboards now have their server set to null when the smash server is deleted

The GameManager now has a isalive function for players, checking if they are contained in the lives hashmap for their server

The kit menu now adjusts its size based on the number of kits

The boss gamemode now shows your personal damage to the boss after the game ends

All players now have at least 0 damage placed into the boss damage hashmap

Gamemodes now control what lobby items are set

getRandomRespawnPoint now tries to place you closest to teammates as a priority

Adds an after game ended function for gamemodes

The testing gamemode now adds lives to players every single update tick so people can join midway through

SmashMaps now throw exceptions for UUID collisions

Fixes a runtime in death when players would leave after dying by using getLives instead of lives

Reduces fish flurry projectile size

borderKill now only teleports if you actually lost a life

Adds an energy flight attribute where you can fly as long as you have energy, and you only regain flight when you touch the ground after stopping flying preventing you from running too much

Scoreboards have now been swapped to a teams system. Instead of setting line values with scores and clearing them every single time we build the scoreboard, we simply assign a team with a unique name based on the score number to that specific score number on the scoreboard. We change the prefix and suffix of that team to update the scoreboard and the prefix and suffix are only updated when they would actually change as well. This should reduce or stop scoreboard flickering entirely.

Scoreboards are now set to update every tick in the SmashServer run so we don't have to call buildscoreboard whenever anything would change, it will detect changes on its own with the new system.

+++ ssm/managers/maps/SmashMap.java
@@ -28,6 +28,9 @@ public abstract class SmashMap implements Listener {
     public SmashMap(File original_directory) {
         Bukkit.getServer().getPluginManager().registerEvents(this, Main.getInstance());
         File copy_directory = new File("maps/_Copies/" + UUID.randomUUID());
+        if(copy_directory.exists()) {
+            throw new RuntimeException("UUID Collision for Mapfile");
+        }
         if (!copy_directory.exists()) {
             try {
                 FileUtils.copyDirectory(original_directory, copy_directory);
